+++ accounting/RELEASES.md
@@ -39,3 +39,13 @@ Based on:
 - [typescript v0.4.1] accounting
 ### Releases
 - [NPM v0.4.1] https://www.npmjs.com/package/@speakeasy-sdks/accounting/v/0.4.1 - accounting
+
+## 2023-11-27 01:49:20
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Accounting.yaml
+- Speakeasy CLI 1.122.0 (2.195.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.4.2] accounting
+### Releases
+- [NPM v0.4.2] https://www.npmjs.com/package/@speakeasy-sdks/accounting/v/0.4.2 - accounting